+++ src/components/datagrid/datagrid.editors.js
@@ -502,7 +502,11 @@ const editors = {
 
     this.init = function () {
       this.input = $('<input class="datepicker" autocomplete="off"/>').appendTo(container);
-      this.input.datepicker(column.editorOptions || { dateFormat: column.dateFormat });
+      let options = column.editorOptions;
+      if (!options && column.dateFormat) {
+        options = { dateFormat: column.dateFormat };
+      }
+      this.input.datepicker(options);
     };
 
     this.val = function (v) {