+++ src/utils/utils.js
@@ -778,7 +778,7 @@ utils.getHiddenSize = function getHiddenSize(el, options) {
   }
 
   el = $(el);
-  options = $.extend({}, defaults, options);
+  options = { ...defaults, ...options };
 
   // element becomes clone and appended to a parentElement, if defined
   const hasDefinedParentElement = DOM.isElement(options.parentElement);
@@ -948,13 +948,11 @@ utils.mergeSettings = function mergeSettings(element, incomingOptions, defaultOp
     }
   }
 
-  // Actually get ready to merge incoming options if we get to this point.
-  return utils.extend(
-    true, {},
-    resolveFunctionBasedSettings(defaultOptions || {}),
-    resolveFunctionBasedSettings(incomingOptions),
-    (element !== undefined ? utils.parseSettings(element) : {})
-  ); // possible to run this without an element present -- will simply skip this part
+  return {
+    ...resolveFunctionBasedSettings(defaultOptions || {}),
+    ...resolveFunctionBasedSettings(incomingOptions),
+    ...(element !== undefined ? utils.parseSettings(element) : {})
+  };
 };
 
 /**
@@ -1106,12 +1104,22 @@ utils.getScrollbarWidth = function () {
  * @returns {array|object} The copied array or object.
  */
 utils.deepCopy = function (arrayOrObject) {
+  const references = new Map();
   const copy = (input) => {
     if (typeof input !== 'object' || input === null) {
       return input; // Return the value if input is not an object
     }
+
+    // If an object has already been cloned then return a
+    // reference to that clone to avoid an infinite loop
+    if (references.has(input) === true) {
+      return references.get(input);
+    }
+
     // Create an array or object to hold the values
     const output = Array.isArray(input) ? [] : {};
+    references.set(arrayOrObject, input);
+
     Object.keys(input).forEach((key) => {
       const value = input[key];
       // Recursively (deep) copy for nested objects, including arrays